+++ lib/src/hooks.dart
@@ -16,4 +16,3 @@ part 'tab_controller.dart';
 part 'text_controller.dart';
 part 'focus.dart';
 part 'scroll_controller.dart';
-part 'is_mounted.dart';
